refactor(gql): update workspace-related queries and mutations to use consistent naming conventions and add new provider queries
