+++ web/gql/mutations/providers.ts
@@ -1,0 +1,25 @@
+import { gql } from '@drivebase/sdk';
+
+export const GET_AUTH_URL = gql(`
+  mutation GetAuthUrl($input: GetAuthUrlInput!) {
+    getAuthUrl(input: $input) {
+      url
+    }
+  }
+`);
+
+export const AUTHORIZE_API_KEY = gql(`
+  mutation AuthorizeApiKey($input: AuthorizeApiKeyInput!) {
+    authorizeApiKey(input: $input) {
+      id
+    }
+  }
+`);
+
+export const HANDLE_OAUTH_CALLBACK = gql(`
+  mutation HandleOAuthCallback($input: HandleOAuthCallbackInput!) {
+    handleOAuthCallback(input: $input) {
+      id
+    }
+  }
+`);
