+++ web/gql/queries/workspaces.ts
